refactor: Extract character animation and movement to engine package
Create internal/engine/character package with:
- types.go: Player struct, CompositeFrame, TerrainQuery interface
- animation.go: UpdateAnimation for ACT-based frame advancement
- movement.go: UpdateMovement for click-to-move with terrain queries

MapViewer now delegates to character package functions and implements
TerrainQuery interface for walkability and height checks.

Reduces map_viewer.go by ~128 lines (3611 → 3483).
